Addressed issues with project-card content alignment.
Once I started adding content to the project-cards, the alignment of the items within were ruined. For example, the modal-trigger would not stick to the bottom of the project-card (because the project-summary div was not stretching to fill the extra space).
I turned both the project-card and project-summary into column-based flexboxes and added margin-auto to the modal-trigger.
I took .modal-overlay out of .project-summary to make project-summary easier to work with and look at. I adjusted the DOM-relative modal-trigger script accordingly.
Adjusted width of modal-title at all screen sizes.