Added Cascade project content, updated accordion scripts.
Made many adjustments to styles for project-card content based on appearance of Cascade Manual content.
Moved accordion-trigger styles to a mixin so it could be resused.
Added a "Results" button and accordion to Cascade Manual project-card.
Refactored scripts for accordion.
Added FontAwesome caret and animation for rotating it when clicked.