Began switching out project-info and project-summary.
I realized that the items I had in project-info were more at-a-glance and should really be the stuff on the project card, while the project-summary content required more reading and should be in the modal.
I also wrote notes to refactor the accordion code, since I will soon need multiple accordions in the same container.