Fixed accordion scripts.
When tested, the previous accordion script was actually opening all accordions of the same class. I had to revise the code to get it to only open the correct content. I had to loop through the content instead of the triggers.
Also added summary-trigger to each project-card.
Added fixed pixel-based font-size to accordion-trigger mixin.