Made $color-accent brighter. Fixed alert close animation.
Used the $warning-color I was using for the alert as the new $color-accent, since the dark green I had was too dark.

I added setTimeout to the close alert callback function within alertMessage(). This finally allows me to also have a slideUp() jQuery animation before the alert is removed from the DOM.